Move oci_distribution to internal oci package
